+++ R/build_schedule.R
@@ -1,19 +1,14 @@
-#' Build a schedule table
+#' Build a schedule
 #'
-#' Builds a schedule data.frame for scheduling pipelines in `run_schedule()`.
+#' Builds a `MaestroSchedule` object for use in `run_schedule()`.
 #'
 #' This function parses the maestro tags of functions located in `pipeline_dir` which is
-#' conventionally called 'pipelines'. An orchestrator requires a schedule table
-#' to determine which pipelines are to run and when. Each row in a schedule table
-#' is a pipeline name and its scheduling parameters such as its frequency.
+#' conventionally called 'pipelines'. An orchestrator requires a `MaestroSchedule`
+#' to determine which pipelines are to run and when. Each pipeline in the schedule
+#' is a parsed function and its scheduling parameters such as its frequency.
 #'
-#' The schedule table is mostly intended to be used by `run_schedule()` immediately.
+#' The `MaestroSchedule` is mostly intended to be passed directly to `run_schedule()`.
 #' In other words, it is not recommended to make changes to it.
-#' 
-#' It is recommended to build the schedule from scratch on each run of the orchestrator
-#' script rather than reusing or caching the schedule object. This is because the 
-#' schedule object precomputes and stores a limited set of future run times for each
-#' pipeline.
 #'
 #' @param pipeline_dir path to directory containing the pipeline scripts
 #' @param quiet silence metrics to the console (default = `FALSE`)
@@ -44,7 +39,7 @@ build_schedule <- function(pipeline_dir = "./pipelines", quiet = FALSE) {
   )
 
   # Error if the directory has no .R scripts
-  if(length(pipelines) == 0) {
+  if (length(pipelines) == 0) {
     cli::cli_inform(
       "No R scripts in {pipeline_dir}."
     )