+++ R/suggest_orch_frequency.R
@@ -9,11 +9,15 @@
 #' Note this function is intended to be used interactively when deciding how often to
 #' schedule the orchestrator. Programmatic use is not recommended.
 #'
+#' **Deprecated.** This function is no longer actively maintained and may be removed
+#' in a future release.
 #' @param schedule MaestroSchedule object created by `build_schedule()`
 #' @inheritParams get_pipeline_run_sequence
 #'
 #' @return frequency string
 #' @export
+#'
+#' @keywords internal
 #'
 #' @examples
 #'
@@ -24,6 +28,8 @@
 #'   suggest_orch_frequency(schedule)
 #' }
 suggest_orch_frequency <- function(schedule, check_datetime = lubridate::now(tzone = "UTC")) {
+
+  lifecycle::deprecate_warn("1.1.0", "suggest_orch_frequency()")
 
   # Check that schedule is a MaestroSchedule
   if (!"MaestroSchedule" %in% class(schedule)) {
@@ -52,7 +58,6 @@ suggest_orch_frequency <- function(schedule, check_datetime = lubridate::now(tzo
     purrr::possibly(convert_to_seconds, otherwise = NA, quiet = TRUE)
   )
 
-  # If the minimum schedule seconds is lte 15 minutes, return the corresponding frequency
   if (min(sch_secs, na.rm = TRUE) <= (60 * 15)) {
     return(schedule$frequency[[which.min(sch_secs)]])
   }