+++ data-raw/maestro_tags.R
@@ -1,9 +1,0 @@
-maestro_tags <- c(
-  "maestroFrequency",
-  "maestroStartTime",
-  "maestroTz",
-  "maestroSkip",
-  "maestroLogLevel"
-)
-
-usethis::use_data(maestro_tags, overwrite = TRUE)