+++ tests/testthat/test_pipelines/test_pipeline_daily_single_bad.R
@@ -1,9 +1,7 @@
 #' Simple mtcars print function
 #'
 #' @maestroFrequency minutely
-#' @maestroStartTime 2024-03-01 09:00:00
 #' @maestroTz UTC
-#'
 #'
 #' @export
 get_mtcars <- function() {