+++ tests/testthat/test_pipelines/test_pipeline_weekday_wrong_freq.R
@@ -1,0 +1,5 @@
+#' @maestroFrequency 1 month
+#' @maestroStartTime Mon 04:00:00
+my_bad_pipeline <- function() {
+  TRUE
+}