+++ tests/testthat/test_pipelines/test_pipeline_weekly_weekday.R
@@ -1,0 +1,5 @@
+#' @maestroFrequency 1 week
+#' @maestroStartTime Mon 04:00:00
+my_weekly_pipeline <- function() {
+  TRUE
+}